refactor(tests): move public-API mcp_server tests from inline mod to tests/mcp_server_*.rs
Migrates 70 tests out of `src/mcp/server.rs::mod tests` into 6 per-tool-group
integration test files under `tests/`, leaving 52 white-box tests inline
with a module-level annotation listing the private items they exercise.
Pure mechanical refactor: no production code, visibility, or test logic
changes — only test relocation with adjusted `use` paths.

New files (each ≤ 1500-LOC budget):
- tests/mcp_server_get_put.rs (11 tests, 301 LOC) — memory_get/memory_put
  validation, ambiguity, OCC conflicts, oversized/empty rejection, and
  the put_from_string variant guard.
- tests/mcp_server_query_search_list.rs (12 tests, 444 LOC) — read-side
  tools: hybrid query, FTS5 sanitization, collection filtering, write-target
  defaults, and the depth=auto contract.
- tests/mcp_server_link_graph.rs (13 tests, 384 LOC) — link/link_close,
  backlinks (limits + temporal), and depth-bounded graph traversal.
- tests/mcp_server_check_timeline_tags.rs (11 tests, 297 LOC) — memory_check
  contradiction surfaces, timeline rendering, and tag list/add/remove.
- tests/mcp_server_gap_stats.rs (16 tests, 377 LOC) — memory_gap/gaps
  redaction + idempotency + boundaries, memory_stats schema, and
  memory_raw input validation.
- tests/mcp_server_misc.rs (7 tests, 232 LOC) — get_info capability,
  memory_add_turn validation paths, memory_close_action status validation.

Shared helpers extracted to tests/common/mcp_harness.rs (112 LOC):
`open_test_db`, `create_page`, `create_page_in_collection`,
`insert_collection`, `set_collection_state`, `extract_text` — all built
exclusively on the public `quaid::mcp::server` and `quaid::core::db`
surfaces, no visibility widened.

Inline residue (52 tests, ~1997 LOC):
private items exercised include `extraction_enabled`,
`extraction_debounce_ms`, `serialize_response`, `validate_slug`,
`validate_relationship`, `validate_tag_list`, `validate_temporal_value`,
`parse_temporal_filter`, `map_graph_error`, `memory_close_action_impl`,
`canonical_slug`, `map_anyhow_error`, `MAX_SLUG_LEN`, `MAX_TAGS_PER_REQUEST`,
`TemporalFilter`, `vault_sync::set_collection_recovery_in_progress_for_test`,
and the private `QuaidServer::db` field used for state-verification queries.

Verification:
  baseline (parent f0e623e):  passed=1357 failed=0
  this commit:                passed=1357 failed=0

Both clippy channels green:
  cargo clippy --all-targets --locked -- -D warnings
  cargo clippy --all-targets --no-default-features \
    --features bundled,online-model --locked -- -D warnings

src/mcp/server.rs:  5926 LOC -> 4022 LOC

Part of OpenSpec change `extract-inline-tests-to-integration`.
